1.8.1
Remade the sidebar and the buffs back to WorldUpdate because timers are weird and didn't work sometimes.
Made mounts actually get put in the misc equipment section.
Put the lavarise counter above the score so that you can still see it at all times.
You can now set games to not be joinable
Fixed the game not using all the accessories (those locked behind expert and master mode)
added gimmicks:
golf cart
feral bite because i'm feeling evil
and a message for a honey arena